Update setup.py
I cant seem to get setup.py to work without having the auth_plugin='mysql_native_password'. When its not there I get an error about SHA2 hashing. It might be because of the sql version im using but im not sure. And then I get another error when using with and as statements so I tried to use try and finally statements and that seemed to work for me.